Prototype 5
Fixed the camera problem i was having with the camera was not following the player.
Created a field for player to get collectables. Have not made the walls move and collectables can not be collected yet.
Made enemies spawn from a cloud. Problem: the enemies only spawn in one spot.
Made a goal for the player to get to. Have not yet implement the goal to give the player a end goal.